requirements.txt
asgiref==3.2.10
Babel==2.8.0
beautifulsoup4==4.9.1
Django==3.0.8
django-bootstrap4==2.2.0
django-countries==6.1.2
django-icons==2.1.1
django-mapbox-location-field==1.6.3
django-phone-field==1.8.1
django-phonenumber-field==4.0.0
freeze==3.0
phonenumbers==8.12.7
Pillow==7.2.0
pytz==2020.1
six==1.15.0
sorl-thumbnail==12.6.3
soupsieve==2.0.1
sqlparse==0.3.1
